requirements.txt
bootstrap-admin==0.3.6
Django==1.8
django-bootstrap-form==3.2
django-bootstrap3==6.2.2
django-bootstrap3-datepicker==0.3.1
django-datetime-widget==0.9.3
MySQL-python==1.2.5
pytz==2015.7
wheel==0.24.0
xlrd==0.9.4
XlsxWriter==0.8.4
